LF-4380 Consolidate middleware for reuse in edit endpoint
