+++ packages/api/src/routes/animalRoute.js
@@ -19,23 +19,20 @@ const router = express.Router();
 import checkScope from '../middleware/acl/checkScope.js';
 import hasFarmAccess from '../middleware/acl/hasFarmAccess.js';
 import AnimalController from '../controllers/animalController.js';
-import AnimalModel from '../models/animalModel.js';
-import {
-  checkAnimalEntities,
-  validateAnimalBatchCreationBody,
-} from '../middleware/checkAnimalEntities.js';
 import multerDiskUpload from '../util/fileUpload.js';
 import validateFileExtension from '../middleware/validation/uploadImage.js';
 import {
   checkRemoveAnimalOrBatch,
   checkEditAnimalOrBatch,
+  checkCreateAnimalOrBatch,
+  checkDeleteAnimalOrBatch,
 } from '../middleware/validation/checkAnimalOrBatch.js';
 
 router.get('/', checkScope(['get:animals']), AnimalController.getFarmAnimals());
 router.post(
   '/',
   checkScope(['add:animals']),
-  validateAnimalBatchCreationBody(),
+  checkCreateAnimalOrBatch('animal'),
   AnimalController.addAnimals(),
 );
 router.patch(
@@ -55,7 +52,7 @@ router.patch(
 router.delete(
   '/',
   checkScope(['delete:animals']),
-  checkAnimalEntities(AnimalModel),
+  checkDeleteAnimalOrBatch('animal'),
   AnimalController.deleteAnimals(),
 );
 router.post(
